+++ src/performance2.spec.js
@@ -1,7 +1,7 @@
 /* global test, jasmine */
 import execa from 'execa'
 import {t} from 'jest-t-assert'
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 25e3 // eslint-disable-line fp/no-mutation
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 40e3 // eslint-disable-line fp/no-mutation
 
 global.Promise = require.requireActual(`bluebird`) // eslint-disable-line fp/no-mutation
 const cwd = process.cwd()
@@ -14,7 +14,7 @@ test(`katsu-curry/debug.curry is slow but useful`, (done) => {
         (output) => {
           const lines = output.stdout.split(`\n`)
           console.log(output.stdout)
-          t.is(lines[3], `Fastest is katsu-curry.curry`)
+          t.is(lines[lines.length - 1], `Fastest is lodash.curry`)
           resolve()
           done()
         }
