--- src/performance3.spec.js
@@ -1,26 +1,0 @@
-/* global test, jasmine */
-import execa from 'execa'
-import {t} from 'jest-t-assert'
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 25e3 // eslint-disable-line fp/no-mutation
-
-global.Promise = require.requireActual(`bluebird`) // eslint-disable-line fp/no-mutation
-const cwd = process.cwd()
-
-test(`katsu-curry could be faster`, (done) => {
-  t.plan(1)
-  return new global.Promise((resolve, reject) => {
-    setImmediate(() => {
-      execa.shell(`node ${cwd}/src/performance3.fixture.js`).then(
-        (output) => {
-          const lines = output.stdout.split(`\n`)
-          console.log(output.stdout)
-          t.is(lines[3], `Fastest is lodash`)
-          resolve()
-          done()
-        }
-      ).catch(
-        reject
-      )
-    })
-  })
-})
